complex: use factory, abstract and interface for CalculatorComplex classes
Therefore, new Calculators can be created by implementing "only 1 file" and the rest of classes will handle it the same way. This is est for people developing different and,or, etc calculations, they just need to focus on the algorithm, and implement in one of the 6 methods (and,or,counter for prop and opponent 3*2=6), and nothing else.